refactor: useGetSecondary 훅, 공통훅으로 분리
