+++ src/constants/@queryKeys/index.ts
@@ -2,6 +2,7 @@ export const COMMON_KEYS = Object.freeze({
   CATEGORY: ['category'],
   USER_PROFILE: (memberId: number) => ['user_profile', memberId],
   RECOMMENDED_USERS: ['recommended-users'],
+  SECONDARY_TOKEN: (type: 'chat' | 'alarm') => ['secondary-token', type],
 });
 
 export const USER_KEYS = Object.freeze({
@@ -13,7 +14,6 @@ export const USER_KEYS = Object.freeze({
 });
 
 export const CHAT_KEYS = Object.freeze({
-  SECONDARY_TOKEN: (type: 'chat' | 'alarm') => ['secondary-token', type],
   CHAT_LIST: ['chat-list'],
   CHAT_DETAIL: (roomId: number) => ['chat-detail', roomId],
   CHAT_MEMBER_LIST: (roomId: number) => ['chat-member-list', roomId],
