+++ src/hooks/@common/useGetSecondaryToken.ts
@@ -1,5 +1,5 @@
 import { getSecondaryToken } from '@/apis/chat';
-import { CHAT_KEYS } from '@/constants/@queryKeys';
+import { COMMON_KEYS } from '@/constants/@queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
 interface IUseGetSecondaryToken {
@@ -12,7 +12,7 @@ export const useGetSecondaryToken = ({
   isLoggedIn,
 }: IUseGetSecondaryToken) => {
   return useQuery({
-    queryKey: CHAT_KEYS.SECONDARY_TOKEN(type),
+    queryKey: COMMON_KEYS.SECONDARY_TOKEN(type),
     queryFn: getSecondaryToken,
     enabled: !!isLoggedIn,
   });
